requirements.txt
mecab-python3==1.0.1
oauthlib==3.1.0
requests==2.24.0
requests-oauthlib==1.3.0
numpy==1.18.5   